requirements.txt
accelerate==1.13.0
addict==2.4.0
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
babel==2.18.0
blis==1.3.3
brotli==1.2.0
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
cmake==4.2.3
colorama==0.4.6
confection==0.1.5
csvw==3.7.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.13
Cython==3.2.4
dlinfo==2.0.0
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
espeakng-loader==0.2.4
exceptiongroup==1.3.1
fastapi==0.135.1
ffmpeg-python==0.2.0
ffmpy==1.0.0
filelock==3.25.2
fsspec==2026.2.0
future==1.0.0
gradio==6.9.0
gradio_client==2.3.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kokoro==0.9.4
language-tags==1.2.0
llvmlite==0.46.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
misaki==0.9.4
more-itertools==10.8.0
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.4.2
ninja==1.13.0
num2words==0.5.14
numba==0.64.0
numpy==2.2.6
openai-whisper==20250625
orjson==3.11.7
packaging==26.0
pandas==2.3.3
phonemizer-fork==3.3.2
pillow==12.1.1
preshed==3.0.12
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
rdflib==7.6.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rfc3986==1.5.0
rich==14.3.3
rpds-py==0.30.0
safehttpx==0.1.7
safetensors==0.7.0
scipy==1.15.3
segments==2.4.0
semantic-version==2.10.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
soundfile==0.13.1
spacy==3.8.11
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.52.1
sympy==1.14.0
termcolor==3.3.0
thinc==8.3.10
tiktoken==0.12.0
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.10.0
torchaudio==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
wasabi==1.1.3
weasel==0.4.3
win32_setctime==1.2.0
wrapt==2.1.2
yt-dlp==2026.3.17
